requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
python-dotenv
pymysql
qrcode
cryptography
Werkzeug
pytest
